visiOCR/requirements.txt
Django==3.2.9
mysql-connector-python==8.0.27
pytesseract==0.3.8
opencv-python==4.5.3.56
xhtml2pdf==0.2.5
qrcode[pil]==7.3.1
Pillow==8.3.1
